+++ station-bundle-impl/src/main/java/com/taobao/cun/auge/punish/bo/impl/PartnerPunishBoImpl.java
@@ -87,9 +87,9 @@ public class PartnerPunishBoImpl implements PartnerPunishBo {
                 if(punishValue.getIndicateName().equals("自然年总违规次数")){
                     vpInfoDto.setTotalIllegalCount(Integer.valueOf(punishValue.getValue().toString()));
                 }else if(punishValue.getIndicateName().equals("严重违规总分数")){
-                    vpInfoDto.setSeriousViolationPoints(Double.valueOf(punishValue.getValue().toString()));
+                    vpInfoDto.setSeriousViolationPoints(Double.valueOf(punishValue.getValue().toString()).intValue());
                 }else if(punishValue.getIndicateName().equals("当前一般违规总分数")){
-                    vpInfoDto.setGeneralViolationPoints(Double.valueOf(punishValue.getValue().toString()));
+                    vpInfoDto.setGeneralViolationPoints(Double.valueOf(punishValue.getValue().toString()).intValue());
                 }
             }
